+++ __tests__/formatMoney.test.ts
@@ -19,7 +19,9 @@ describe('formatMoney', () => {
   test("using options", () => {
     expect(formatMoney(1.532, { precision: 2, currency: "PLN" }, "pl-PL")).toEqual(`1,53${IntlWhitespace}zł`);
     expect(formatMoney(1.532, { precision: 2, currency: "USD" }, "pl-PL")).toEqual(`1,53${IntlWhitespace}USD`);
+    expect(formatMoney(1.532, { precision: 2, currency: 840 }, "pl-PL")).toEqual(`1,53${IntlWhitespace}USD`);
     expect(formatMoney(1.53588, { precision: 3, currency: "USD" }, "en")).toEqual("$1.536");
+    expect(formatMoney(-1.53588, { precision: 3, currency: "USD", currencySign: "accounting" }, "en")).toEqual("($1.536)");
     expect(formatMoney(1.53588, { precision: 3, currency: "USD", currencyDisplay: "code" }, "en")).toEqual(`USD${IntlWhitespace}1.536`);
   });
 
@@ -27,6 +29,10 @@ describe('formatMoney', () => {
     expect(formatMoney(1.53588, { }, "en")).toEqual("1.54");
     expect(formatMoney(1.53588, { precision: 3 }, "en")).toEqual("1.536");
     expect(formatMoney(1.53588, { currency: "USD" }, "en")).toEqual("$1.54");
+    expect(formatMoney(1.53588, { currency: "CAD" }, "en")).toEqual("CA$1.54");
+    expect(formatMoney(1.53588, { currency: "CAD", currencyDisplay: "narrowSymbol" }, "en")).toEqual("$1.54");
+    expect(formatMoney(1.53588, { currencyDisplay: "narrowSymbol" }, "en")).toEqual("1.54");
+    expect(formatMoney(-1.53588, { currencySign: "accounting" }, "en")).toEqual("-1.54");
     expect(formatMoney(1.53588, { currency: "USD", currencyDisplay: "code" }, "en")).toEqual(`USD${IntlWhitespace}1.54`);
   });
 });