+++ src/types/intlCurrencySign.ts
@@ -1,0 +1,1 @@
+export type IntlCurrencySign = "standard" | "accounting";