+++ src/types/numberFormatMoney.ts
@@ -1,8 +1,10 @@
 import { IntlCurrencyDisplay } from "./intlCurrencyDisplay";
-import { Iso4217Alpha3Code } from "./iso4217";
+import { IntlCurrencySign } from "./intlCurrencySign";
+import { Iso4217Alpha3Code, Iso4217NumericCode } from "./iso4217";
 
 export interface NumberFormatMoney {
   precision: number;
-  currency: Iso4217Alpha3Code;
+  currency: Iso4217Alpha3Code | Iso4217NumericCode;
   currencyDisplay: IntlCurrencyDisplay;
+  currencySign: IntlCurrencySign;
 }